# SALTO-843: Improve graph operations performance