--- a/packages/dag/src/diff.ts
+++ b/packages/dag/src/diff.ts
@@ -14,7 +14,7 @@
 * limitations under the License.
 */
 import _ from 'lodash'
-import wu, { WuIterable } from 'wu'
+import wu from 'wu'
 import { collections } from '@salto-io/lowerdash'
 import { NodeId, DataNodeMap } from './nodemap'
 
@@ -82,18 +82,15 @@ const mergeNodes = <T>(
   target: DiffGraph<T>, oldIds: NodeId[], newId: NodeId, newData: DiffNode<T>
 ): void => {
   const deps = new Set<NodeId>(
-    wu.chain(oldIds.map(id => target.get(id)))
-      .flatten()
-      // filter out old nodes from dependecy list
+    wu.chain(...oldIds.map(id => target.get(id)))
+      // filter out old nodes from dependency list
       .filter(id => !oldIds.includes(id))
   )
 
-  // update reverse deps to new node
-  oldIds.forEach(
-    oldId => target.deleteNode(oldId).forEach(affected => target.get(affected).add(newId))
-  )
-
   target.addNode(newId, deps, newData)
+
+  // update reverse deps to new node
+  target.deleteNode(...oldIds).forEach(affected => target.get(affected).add(newId))
 }
 
 const tryCreateModificationNode = <T>(
@@ -101,10 +98,30 @@ const tryCreateModificationNode = <T>(
   additionNodeId: NodeId,
   removalNodeId: NodeId,
 ): DiffGraph<T> => {
+  const edgeCausesCycle = (from: NodeId, to: NodeId): boolean => {
+    const fromDeps = target.get(from)
+    // no modification required, this is already checked
+    if (fromDeps.has(to)) return false
+    const toDeps = target.get(to)
+    const newEdges = new Map([
+      [from, new Set([...fromDeps, to])],
+      // Remove edge in the reverse direction to avoid a false positive cycle
+      [to, new Set(wu(toDeps).filter(dep => dep !== from))],
+    ])
+    return target.doesCreateCycle(newEdges, from)
+  }
+  // for cycle detection, merging the nodes is equivalent to adding edges in both directions
+  // (obviously not at the same time, because that would create a cycle)
+  if (edgeCausesCycle(additionNodeId, removalNodeId)
+      || edgeCausesCycle(removalNodeId, additionNodeId)) {
+    return target
+  }
+
   const removalNode = target.getData(removalNodeId) as RemovalDiffNode<T>
   const additionNode = target.getData(additionNodeId) as AdditionDiffNode<T>
   const { originalId } = removalNode
 
+  const modificationNodeId = _.uniqueId()
   const modificationNode: ModificationDiffNode<T> = {
     action: 'modify',
     originalId,
@@ -114,16 +131,17 @@ const tryCreateModificationNode = <T>(
     },
   }
 
-  const modificationNodeId = _.uniqueId()
-
-  return target.tryTransform(t => {
-    mergeNodes(t, [removalNodeId, additionNodeId], modificationNodeId, modificationNode)
-    return modificationNodeId
-  })[0]
+  mergeNodes(target, [removalNodeId, additionNodeId], modificationNodeId, modificationNode)
+  return target
 }
 
-type SetIdPair = [collections.set.SetId, collections.set.SetId]
-export const mergeNodesToModify = async <T>(target: DiffGraph<T>): Promise<DiffGraph<T>> => {
+type Pair<T> = [T, T]
+const isPair = <T>(items: ReadonlyArray<T>): items is Pair<T> => (
+  items.length === 2
+)
+type SetIdPair = Pair<collections.set.SetId>
+
+export const mergeNodesToModify = <T>(target: DiffGraph<T>): void => {
   const addBeforeRemove = (nodes: SetIdPair): SetIdPair => {
     const [adds, removes] = _.partition(nodes, node => target.getData(node).action === 'add')
     return [adds[0], removes[0]]
@@ -132,9 +150,9 @@ export const mergeNodesToModify = async <T>(target: DiffGraph<T>): Promise<DiffG
   // Find all pairs of nodes pointing to the same original ID
   const mergeCandidates = wu(
     iterable.groupBy(target.keys(), id => target.getData(id).originalId).values()
-  ).filter(nodes => nodes.length === 2) as WuIterable<SetIdPair>
+  ).filter(isPair)
 
-  return mergeCandidates
+  mergeCandidates
     .map(addBeforeRemove)
     .reduce(
       (graph, [add, remove]) => tryCreateModificationNode(graph, add, remove),