General changes:
- Avoid cloning graphs for each node
- Remove multiple nodes to slightly improve deleteNode efficiency (still O(n))

mergeNodesToModify:
- check for cycles before creating the merged node

Group graph:
- Add index for node id -> group id to avoid O(n) searches for each node
- add items to existing node instead of creating a new one and merging
- removed log that wasn't very useful and took a non-trivial percentage of time

---

Based on top of #1225 - please review separately (only review second commit here)

profiling:

scenario - vanilla SFDC workspace, for each element add an annotation / value to create a modification diff, for each type create a copy new type to create an addition diff and rename all fields from the existing type to create remove and add changes that would be merged to the same group

before:
![image](https://user-images.githubusercontent.com/1883898/87668278-80d81500-c774-11ea-8546-2a32e7dbbc12.png)

after:
![image](https://user-images.githubusercontent.com/1883898/87668339-98af9900-c774-11ea-98b8-d5e17602fe17.png)

| function | before (ms) | after (ms) |
|---|---|---|
| updateDeps | 3220 | 0 |
| filterInvalidChanges | 1540 | 0 |
| mergeNodesToModify | 3540 | 351 |
| buildGroupedGraph... | 5160 | 2100 |

note that after the changes most (2000 out of 2100) of the time taken by buildGroupedGraph is due to the "walk" (which is mostly due to the time it takes to `deleteNode`), so the next PR will focus on that part